+++ RBSDKPlayer.framework/Headers/RBSDKPlayerMediaInfo.h
@@ -28,11 +28,6 @@ NS_ASSUME_NONNULL_BEGIN
 @property (weak, nonatomic, nullable) NSString *caption;
 
 /**
- Angle position index
- */
-@property (nonatomic) NSInteger anglePosition;
-
-/**
  Dictionary with other useful information.
  e.g. referenceId
  */
@@ -45,12 +40,11 @@ NS_ASSUME_NONNULL_BEGIN
 
  @param mediaId The media id
  @param caption The media's caption (title)
- @param anglePosition The media's angle position (on the horizontal scrolling)
  @param otherParameters Other useful info e.g. referenceId
 
  @return An instance of `RBSDKPlayerMediaInfo`
  */
-- (instancetype)initWithMediaId:(nullable NSString *)mediaId caption:(nullable NSString *)caption anglePosition:(NSInteger)anglePosition otherParameters:(nullable NSDictionary *)otherParameters;
+- (instancetype)initWithMediaId:(nullable NSString *)mediaId caption:(nullable NSString *)caption otherParameters:(nullable NSDictionary *)otherParameters;
 
 @end