+++ RBSDKPlayer.framework/Headers/RBSDKRhinobirdPlayerViewController.h
@@ -9,6 +9,7 @@
 #import "RBSDKPlayer.h"
 
 @class RBSDKMoment;
+@class RBSDKMomentStream;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -23,22 +24,26 @@ NS_ASSUME_NONNULL_BEGIN
 
  @param momentId Moment's string id
  @param options RBSDKPlayerOption
+ @param delegate Related controller delegate.
 
  @return An instance of `RBSDKPlayerViewController`
  */
 - (nonnull instancetype)initWithMomentId:(nonnull NSString *)momentId
-                                 options:(RBSDKPlayerOption)options;
+                                 options:(RBSDKPlayerOption)options
+                                delegate:(id)delegate;
 
 /**
  Initialize a player controller with a specific moment
 
  @param moment RBSDKMoment New Moment model.
  @param options RBSDKPlayerOption
+ @param delegate Related controller delegate.
 
  @return An instance of `RBSDKPlayerViewController`
  */
 - (nonnull instancetype)initWithNewMoment:(nonnull RBSDKMoment *)moment
-                                  options:(RBSDKPlayerOption)options;
+                                  options:(RBSDKPlayerOption)options
+                                  delegate:(id)delegate;
 
 /**
  Add another moment to the player (enabling the vertical swiping), there isn't a
@@ -49,15 +54,33 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)addMomentId:(nonnull NSString *)momentId;
 
 /**
+ Add another new moment to the player (enabling the vertical swiping).
+ There isn't a limit on the amount of moments that you can add.
+
+ @param newMoment New Moment.
+ */
+- (void)addNewMoment:(RBSDKMoment *)newMoment;
+
+/**
+ Add moment streams to the current moment.
+ There isn't a limit on the amount of moment streans that you can add.
+
+ @param momentStreamsArray Array of moment streams.
+ */
+- (void)addMomentStreamsArray:(NSArray<RBSDKMomentStream *> *)momentStreamsArray;
+
+/**
  Initialize a player controller with a specific collection-id
 
  @param collectionId collection's string id
  @param options RBSDKPlayerOption
+ @param delegate Related controller delegate. 
 
  @return An instance of `RBSDKPlayerViewController`
  */
 - (nonnull instancetype)initWithCollectionId:(nonnull NSString *)collectionId
-                                     options:(RBSDKPlayerOption)options;
+                                     options:(RBSDKPlayerOption)options
+                                     delegate:(id)delegate;
 
 @end